requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
rich>=13.0.0
plyer>=2.1.0
apscheduler>=3.10.0
aiohttp>=3.9.0
cryptography>=41.0.0
